+++ src/Chirp.Razor/Chirp.Razor.csproj
@@ -27,7 +27,6 @@
   </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="../Chirp.Core/Chirp.Core.csproj" />
         <ProjectReference Include="../Chirp.Infrastructure/Chirp.Infrastructure.csproj" />
     </ItemGroup>
