+++ test/Chirp.InfrastructureTest/GlobalUsings.cs
@@ -1,6 +1,6 @@
 ﻿global using Xunit;
 
 global using Microsoft.Data.Sqlite;
-global using Chirp.Core.Data;
+global using Chirp.Infrastructure.Data;
 
 global using Microsoft.EntityFrameworkCore;
